requirements.txt
# API y Web
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.7
strawberry-graphql==0.176.1

# Base de datos
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4

# MongoDB (para consultas directas si es necesario)
pymongo==4.3.3
motor==3.1.2
beanie==1.19.0

# Machine Learning
tensorflow==2.18.0


pandas==2.0.1
joblib==1.2.0

# Mensajería
pika==1.3.1

# Utilidades
python-jose==3.3.0
passlib==1.7.4